fix(game): move fullscreen component to shared
