--- src/entities/game/index.ts
@@ -1,1 +1,0 @@
-export { FullscreenWrapper } from './ui';
