--- src/entities/game/ui/index.ts
@@ -1,1 +1,0 @@
-export { FullscreenWrapper } from './fullscreeen-wrapper';
