+++ src/features/game/ui/index.ts
@@ -1,2 +1,2 @@
 export { EndGameDialog } from './end-game-dialog';
-export { FullscreenButton } from './fullscreen-button';
+export { FullscreenButton } from '../../../shared/ui/components/fullscreen-button';
