+++ src/shared/ui/components/fullscreeen-wrapper/fullscreen-wrapper.tsx
