+++ src/shared/ui/components/fullscreen-button/fullscreen-button.tsx
