+++ src/shared/ui/components/fullscreen-button/index.ts
