+++ src/shared/ui/components/index.ts
@@ -1,2 +1,4 @@
+export { FullscreenButton } from './fullscreen-button';
 export { InputField } from './input-field';
 export { getNumberFormatCustom } from './number-format-custom';
+export { FullscreenWrapper } from './fullscreeen-wrapper';
